+++ ci4/app/Interfaces/DeploymentSpecificationVolume.php
@@ -1,0 +1,13 @@
+<?php namespace App\Interfaces;
+/**
+ * @property string $mount_path
+ * @property string $sub_path
+ * @property int $capacity
+ * @property string $volume_mode
+ * @property string $reclaim_policy
+ * @property string $nfs_server
+ * @property string $nfs_path
+ */
+interface DeploymentSpecificationVolume {
+
+}