+++ ci4/app/Interfaces/DeploymentSpecificationVolumeList.php
@@ -1,0 +1,7 @@
+<?php namespace App\Interfaces;
+/**
+ * @property DeploymentSpecificationVolume[] $values
+ */
+interface DeploymentSpecificationVolumeList {
+
+}