+++ vue/src/core/services/Deploy/models/DeploymentSpecificationVolume.ts
@@ -1,0 +1,14 @@
+/**
+ * Created by ModelParser
+ * Date: 11-02-2025.
+ * Time: 10:36.
+ */
+import {DeploymentSpecificationVolumeDefinition} from './definitions/DeploymentSpecificationVolumeDefinition';
+
+export class DeploymentSpecificationVolume extends DeploymentSpecificationVolumeDefinition {
+
+    constructor(json?: any) {
+        super(json);
+    }
+
+}