+++ vue/src/core/services/Deploy/models/definitions/InitContainerDefinition.ts
@@ -18,6 +18,7 @@ export class InitContainerDefinition extends BaseModel {
     container_image_tag_value?: string;
     container_image_pull_policy?: string;
     include_deployment_environment_variables?: boolean;
+    include_volumes?: boolean;
     init_container_environment_variables?: InitContainerEnvironmentVariable[];
     deployment_specifications?: DeploymentSpecification[];
     id?: number;
@@ -46,6 +47,7 @@ export class InitContainerDefinition extends BaseModel {
             delete this.container_image_tag_value;
             delete this.container_image_pull_policy;
             delete this.include_deployment_environment_variables;
+            delete this.include_volumes;
             delete this.init_container_environment_variables;
             delete this.deployment_specifications;
             delete this.id;
@@ -86,6 +88,9 @@ export class InitContainerDefinition extends BaseModel {
         }
         if (data.include_deployment_environment_variables != null) {
             this.include_deployment_environment_variables = data.include_deployment_environment_variables;
+        }
+        if (data.include_volumes != null) {
+            this.include_volumes = data.include_volumes;
         }
         if (data.init_container_environment_variables != null) {
             this.init_container_environment_variables = data.init_container_environment_variables.map((i: any) => new InitContainerEnvironmentVariable(i));